+++ splunk-otel-android/src/test/java/io/opentelemetry/rum/internal/SessionIdRatioBasedSamplerTest.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.splunk.rum;
+package io.opentelemetry.rum.internal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.Span;
@@ -29,17 +30,16 @@ import io.opentelemetry.sdk.trace.samplers.Sampler;
 import io.opentelemetry.sdk.trace.samplers.SamplingDecision;
 import java.util.Collections;
 import java.util.List;
-import java.util.function.Supplier;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class SessionIdRatioBasedSamplerTest {
+    @Mock SessionId sessionId;
     private static final String HIGH_ID = "00000000000000008fffffffffffffff";
-    private static final Supplier<String> HIGH_ID_SUPPLIER = () -> HIGH_ID;
     private static final String LOW_ID = "00000000000000000000000000000000";
-    private static final Supplier<String> LOW_ID_SAMPLER = () -> LOW_ID;
     private static final IdGenerator idsGenerator = IdGenerator.random();
 
     private final String traceId = idsGenerator.generateTraceId();
@@ -49,7 +49,9 @@ class SessionIdRatioBasedSamplerTest {
 
     @Test
     void samplerDropsHigh() {
-        SessionIdRatioBasedSampler sampler = new SessionIdRatioBasedSampler(0.5, HIGH_ID_SUPPLIER);
+        when(sessionId.getSessionId()).thenReturn(HIGH_ID);
+
+        SessionIdRatioBasedSampler sampler = new SessionIdRatioBasedSampler(0.5, sessionId);
 
         // Sampler drops if TraceIdRatioBasedSampler would drop this sessionId
         assertEquals(shouldSample(sampler), SamplingDecision.DROP);
@@ -58,25 +60,35 @@ class SessionIdRatioBasedSamplerTest {
     @Test
     void samplerKeepsLowestId() {
         // Sampler accepts if TraceIdRatioBasedSampler would accept this sessionId
-        SessionIdRatioBasedSampler sampler = new SessionIdRatioBasedSampler(0.5, LOW_ID_SAMPLER);
+        when(sessionId.getSessionId()).thenReturn(LOW_ID);
+
+        SessionIdRatioBasedSampler sampler = new SessionIdRatioBasedSampler(0.5, sessionId);
         assertEquals(shouldSample(sampler), SamplingDecision.RECORD_AND_SAMPLE);
     }
 
     @Test
     void zeroRatioDropsAll() {
-        SessionIdRatioBasedSampler samplerHigh =
-                new SessionIdRatioBasedSampler(0.0, HIGH_ID_SUPPLIER);
+        when(sessionId.getSessionId()).thenReturn(HIGH_ID);
+
+        SessionIdRatioBasedSampler samplerHigh = new SessionIdRatioBasedSampler(0.0, sessionId);
         assertEquals(shouldSample(samplerHigh), SamplingDecision.DROP);
-        SessionIdRatioBasedSampler samplerLow = new SessionIdRatioBasedSampler(0.0, LOW_ID_SAMPLER);
+
+        when(sessionId.getSessionId()).thenReturn(LOW_ID);
+
+        SessionIdRatioBasedSampler samplerLow = new SessionIdRatioBasedSampler(0.0, sessionId);
         assertEquals(shouldSample(samplerLow), SamplingDecision.DROP);
     }
 
     @Test
     void oneRatioAcceptsAll() {
-        SessionIdRatioBasedSampler samplerHigh =
-                new SessionIdRatioBasedSampler(1.0, HIGH_ID_SUPPLIER);
+        when(sessionId.getSessionId()).thenReturn(HIGH_ID);
+
+        SessionIdRatioBasedSampler samplerHigh = new SessionIdRatioBasedSampler(1.0, sessionId);
         assertEquals(shouldSample(samplerHigh), SamplingDecision.RECORD_AND_SAMPLE);
-        SessionIdRatioBasedSampler samplerLow = new SessionIdRatioBasedSampler(1.0, LOW_ID_SAMPLER);
+
+        when(sessionId.getSessionId()).thenReturn(LOW_ID);
+
+        SessionIdRatioBasedSampler samplerLow = new SessionIdRatioBasedSampler(1.0, sessionId);
         assertEquals(shouldSample(samplerLow), SamplingDecision.RECORD_AND_SAMPLE);
     }
